requirements.txt
pyyaml>=6.0
python-dateutil>=2.8.2
typing-extensions>=4.5.0
